Update text line handling
